requirements-dev.txt

# testing
pytest
pytest-aiohttp
pytest-html
requests
